Split up apiRoutes.js into userRoutes and itemRoutes for organizational purposes.
